#260 fix: 퍼센티지 데이터를 정수타입에서 Double 타입으로 변경

+++ src/main/java/codesquad/fineants/domain/portfolio/Portfolio.java
@@ -115,12 +115,12 @@ public class Portfolio extends BaseEntity {
 	}
 
 	// 포트폴리오 총 손익율 = (포트폴리오 총 손익 / 포트폴리오 총 투자 금액) * 100%
-	public Integer calculateTotalGainRate() {
-		long totalInvestmentAmount = calculateTotalInvestmentAmount();
+	public Double calculateTotalGainRate() {
+		Long totalInvestmentAmount = calculateTotalInvestmentAmount();
 		if (totalInvestmentAmount == 0) {
-			return 0;
+			return 0.0;
 		}
-		return (int)(((double)calculateTotalGain() / (double)totalInvestmentAmount) * 100);
+		return (calculateTotalGain().doubleValue() / totalInvestmentAmount.doubleValue()) * 100;
 	}
 
 	// 포트폴리오 총 투자 금액 = 각 종목들의 구입가들의 합계
@@ -149,15 +149,15 @@ public class Portfolio extends BaseEntity {
 
 	// 포트폴리오 당일 손익율 = (당일 포트폴리오 가치 총합 - 이전 포트폴리오 가치 총합) / 이전 포트폴리오 가치 총합
 	// 단, 이전 포트폴리오가 없는 경우 ((당일 포트폴리오 가치 총합 - 당일 포트폴리오 총 투자 금액) / 당일 포트폴리오 총 투자 금액) * 100%
-	public Integer calculateDailyGainRate(PortfolioGainHistory prevHistory) {
+	public Double calculateDailyGainRate(PortfolioGainHistory prevHistory) {
 		double prevCurrentValuation = prevHistory.getCurrentValuation();
 		if (prevCurrentValuation == 0) {
-			double currentValuation = calculateTotalCurrentValuation();
-			double totalInvestmentAmount = calculateTotalInvestmentAmount();
-			return (int)(((currentValuation - totalInvestmentAmount) / totalInvestmentAmount) * 100);
+			double currentValuation = calculateTotalCurrentValuation().doubleValue();
+			double totalInvestmentAmount = calculateTotalInvestmentAmount().doubleValue();
+			return ((currentValuation - totalInvestmentAmount) / totalInvestmentAmount) * 100;
 		}
 		double currentValuation = calculateTotalCurrentValuation();
-		return (int)(((currentValuation - prevCurrentValuation) / prevCurrentValuation) * 100);
+		return ((currentValuation - prevCurrentValuation) / prevCurrentValuation) * 100;
 	}
 
 	// 포트폴리오 당월 예상 배당금 = 각 종목들에 해당월의 배당금 합계
@@ -196,8 +196,8 @@ public class Portfolio extends BaseEntity {
 	}
 
 	// 최대손실율 = ((예산 - 최대손실금액) / 예산) * 100
-	public Integer calculateMaximumLossRate() {
-		return (int)(((double)(budget - maximumLoss) / (double)budget) * 100);
+	public Double calculateMaximumLossRate() {
+		return ((budget.doubleValue() - maximumLoss.doubleValue()) / budget.doubleValue()) * 100;
 	}
 
 	// 투자대비 연간 배당율 = 포트폴리오 총 연배당금 / 포트폴리오 투자금액 * 100
@@ -244,8 +244,8 @@ public class Portfolio extends BaseEntity {
 	}
 
 	// 목표 수익률 = ((목표 수익 금액 - 예산) / 예산) * 100
-	public Integer calculateTargetReturnRate() {
-		return (int)(((targetGain.doubleValue() - budget.doubleValue()) / budget.doubleValue()) * 100);
+	public Double calculateTargetReturnRate() {
+		return ((targetGain.doubleValue() - budget.doubleValue()) / budget.doubleValue()) * 100;
 	}
 
 	// 총 자산 = 잔고 + 평가금액 합계
